+++ app/containers/message/Content.tsx
@@ -10,7 +10,7 @@ import { SYSTEM_MESSAGE_TYPES_WITH_AUTHOR_NAME, getInfoMessage } from './utils';
 import { themes } from '../../constants/colors';
 import MessageContext from './Context';
 import Encrypted from './Encrypted';
-import { E2E_MESSAGE_TYPE } from '../../lib/encryption/constants';
+import { E2E_MESSAGE_TYPE } from '../../lib/constants';
 import { IMessageContent } from './interfaces';
 import { useTheme } from '../../theme';
