+++ app/containers/message/index.tsx
@@ -6,7 +6,7 @@ import Message from './Message';
 import MessageContext from './Context';
 import debounce from '../../utils/debounce';
 import { SYSTEM_MESSAGES, getMessageTranslation } from './utils';
-import { E2E_MESSAGE_TYPE, E2E_STATUS } from '../../lib/encryption/constants';
+import { E2E_MESSAGE_TYPE, E2E_STATUS } from '../../lib/constants';
 import messagesStatus from '../../constants/messagesStatus';
 import { useTheme, withTheme } from '../../theme';
 import openLink from '../../utils/openLink';
