+++ app/definitions/rest/v1/users.ts
@@ -49,6 +49,9 @@ export type UsersEndpoints = {
 	'users.resetAvatar': {
 		POST: (params: { userId: string }) => {};
 	};
+	'users.removeOtherTokens': {
+		POST: (params: { userId: string }) => {};
+	};
 	'users.getPreferences': {
 		GET: (params: { userId: IUser['_id'] }) => {
 			preferences: INotificationPreferences;
