+++ app/lib/methods/actions.ts
@@ -1,18 +1,10 @@
-import random from '../../utils/random';
+import { ITriggerAction, IUserInteraction, ModalActions } from '../../containers/UIKit/interfaces';
+import { TRocketChat } from '../../definitions/IRocketChat';
 import EventEmitter from '../../utils/events';
 import fetch from '../../utils/fetch';
+import random from '../../utils/random';
 import Navigation from '../Navigation';
 import sdk from '../rocketchat/services/sdk';
-import {
-	ActionTypes,
-	ITriggerAction,
-	ITriggerBlockAction,
-	ITriggerCancel,
-	ITriggerSubmitView,
-	IUserInteraction,
-	ModalActions
-} from '../../containers/UIKit/interfaces';
-import { TRocketChat } from '../../definitions/IRocketChat';
 
 const triggersId = new Map();
 
@@ -138,19 +130,4 @@ export function triggerAction(
 		}
 		return reject();
 	});
-}
-
-export default function triggerBlockAction(this: TRocketChat, options: ITriggerBlockAction) {
-	return triggerAction.call(this, { type: ActionTypes.ACTION, ...options });
-}
-
-export async function triggerSubmitView(this: TRocketChat, { viewId, ...options }: ITriggerSubmitView) {
-	const result = await triggerAction.call(this, { type: ActionTypes.SUBMIT, viewId, ...options });
-	if (!result || ModalActions.CLOSE === result) {
-		Navigation.back();
-	}
-}
-
-export function triggerCancel(this: TRocketChat, { view, ...options }: ITriggerCancel) {
-	return triggerAction.call(this, { type: ActionTypes.CLOSED, view, ...options });
 }
