+++ app/lib/methods/loadMessagesForRoom.ts
@@ -4,7 +4,7 @@ import { MessageTypeLoad } from '../../constants/messageTypeLoad';
 import { IMessage, TMessageModel } from '../../definitions';
 import log from '../../utils/log';
 import { getMessageById } from '../database/services/Message';
-import roomTypeToApiType, { RoomTypes } from '../rocketchat/methods/roomTypeToApiType';
+import roomTypeToApiType, { RoomTypes } from './roomTypeToApiType';
 import sdk from '../rocketchat/services/sdk';
 import { generateLoadMoreId } from '../utils';
 import updateMessages from './updateMessages';
