+++ app/lib/rocketchat/methods/getUsersPresence.ts
@@ -1,14 +1,14 @@
 import { InteractionManager } from 'react-native';
 import { sanitizedRaw } from '@nozbe/watermelondb/RawRecord';
 
-import { IActiveUsers } from '../../reducers/activeUsers';
-import { compareServerVersion } from '../utils';
-import { store as reduxStore } from '../auxStore';
-import { setActiveUsers } from '../../actions/activeUsers';
-import { setUser } from '../../actions/login';
-import database from '../database';
-import { IRocketChat, IUser } from '../../definitions';
-import sdk from '../rocketchat/services/sdk';
+import { IActiveUsers } from '../../../reducers/activeUsers';
+import { compareServerVersion } from '../../utils';
+import { store as reduxStore } from '../../auxStore';
+import { setActiveUsers } from '../../../actions/activeUsers';
+import { setUser } from '../../../actions/login';
+import database from '../../database';
+import { IRocketChat, IUser } from '../../../definitions';
+import sdk from '../services/sdk';
 
 export function subscribeUsersPresence(this: IRocketChat) {
 	const serverVersion = reduxStore.getState().server.version as string;
