+++ app/lib/rocketchat/methods/logout.ts
@@ -2,17 +2,17 @@ import * as FileSystem from 'expo-file-system';
 import { Rocketchat as RocketchatClient } from '@rocket.chat/sdk';
 import Model from '@nozbe/watermelondb/Model';
 
-import { getDeviceToken } from '../../notifications/push';
-import { extractHostname } from '../../utils/server';
-import { BASIC_AUTH_KEY } from '../../utils/fetch';
-import database, { getDatabase } from '../database';
-import RocketChat from '../rocketchat';
-import { useSsl } from '../../utils/url';
-import log from '../../utils/log';
-import { E2E_PRIVATE_KEY, E2E_PUBLIC_KEY, E2E_RANDOM_PASSWORD_KEY } from '../encryption/constants';
-import UserPreferences from '../userPreferences';
-import { ICertificate, IRocketChat } from '../../definitions';
-import sdk from '../rocketchat/services/sdk';
+import { getDeviceToken } from '../../../notifications/push';
+import { extractHostname } from '../../../utils/server';
+import { BASIC_AUTH_KEY } from '../../../utils/fetch';
+import database, { getDatabase } from '../../database';
+import RocketChat from '..';
+import { useSsl } from '../../../utils/url';
+import log from '../../../utils/log';
+import { E2E_PRIVATE_KEY, E2E_PUBLIC_KEY, E2E_RANDOM_PASSWORD_KEY } from '../../constants';
+import UserPreferences from '../../userPreferences';
+import { ICertificate, IRocketChat } from '../../../definitions';
+import sdk from '../services/sdk';
 
 function removeServerKeys({ server, userId }: { server: string; userId?: string | null }) {
 	UserPreferences.removeItem(`${RocketChat.TOKEN_KEY}-${server}`);
@@ -98,7 +98,7 @@ export async function removeServer({ server }: { server: string }): Promise<void
 	}
 }
 
-export default async function logout(this: IRocketChat, { server }: { server: string }): Promise<void> {
+export async function logout(this: IRocketChat, { server }: { server: string }): Promise<void> {
 	if (this.roomsSub) {
 		this.roomsSub.stop();
 		this.roomsSub = null;
