+++ app/lib/rocketchat/methods/search.ts
@@ -3,8 +3,8 @@ import { Q } from '@nozbe/watermelondb';
 import { sanitizeLikeString } from '../../database/utils';
 import database from '../../database/index';
 import { spotlight } from '../services/restApi';
-import isGroupChat from './isGroupChat';
 import { ISearch, ISearchLocal, SubscriptionType } from '../../../definitions';
+import { isGroupChat } from './helpers';
 
 let debounce: null | ((reason: string) => void) = null;
