+++ app/utils/theme.ts
@@ -6,8 +6,9 @@ import { IThemePreference, TThemeMode } from '../definitions/ITheme';
 import { themes } from '../constants/colors';
 import { isAndroid } from './deviceInfo';
 import UserPreferences from '../lib/userPreferences';
-import { THEME_PREFERENCES_KEY } from '../lib/rocketchat';
+import { THEME_PREFERENCES_KEY } from '../lib/constants';
 import { TSupportedThemes } from '../theme';
+
 
 let themeListener: { remove: () => void } | null;
