+++ app/views/E2ESaveYourPasswordView.tsx
@@ -12,7 +12,7 @@ import StatusBar from '../containers/StatusBar';
 import { LISTENER } from '../containers/Toast';
 import { IApplicationState, IBaseScreen } from '../definitions';
 import I18n from '../i18n';
-import { E2E_RANDOM_PASSWORD_KEY } from '../lib/encryption/constants';
+import { E2E_RANDOM_PASSWORD_KEY } from '../lib/constants';
 import UserPreferences from '../lib/userPreferences';
 import { E2ESaveYourPasswordStackParamList } from '../stacks/types';
 import { withTheme } from '../theme';
